Setup the Database backend using djongo and Mongo
create a db in mongo named healthmasys_db with username healthmasys.
The role of the user healthmasys is dbOwner and it uses password
healthmasysadmin .

Make db migrations using
$ python manage.py makemigrations
and then migrate the data into the db with
$ python manage.py migrate
in your virtualenv

A superuser named healthmasys was created with the same password
as the db. This administrator can be accessed using the django admin page at
127.0.0.1:8000/admin after the runserver command has been executed
to start up the development server